testeos y retoques finales

+++ src/app/features/courses/curso-creator/curso-creator.component.ts
@@ -4,6 +4,9 @@ import { MatTableDataSource } from '@angular/material/table';
 import { Curso } from '../../../shared/interfaces/interfaces';
 import * as CursoActions from '../store/courses.actions';
 import { selectCursos } from '../store/courses.selectors';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-curso-creator',
@@ -20,7 +23,11 @@ export class CursoCreatorComponent implements OnInit {
   nuevoCurso: Curso = { nombre: '', descripcion: '' };
   isEdit = false;
 
-  constructor(private store: Store) {}
+  constructor(
+    private store: Store, 
+    private dialog: MatDialog,
+    private snackBar: MatSnackBar
+  ) {}
 
   ngOnInit(): void {
     this.store.dispatch(CursoActions.cargarCursos());
@@ -32,8 +39,13 @@ export class CursoCreatorComponent implements OnInit {
 
   agregarCurso(cursoForm: any): void {
     if (cursoForm.valid) {
-      if (this.isEdit && this.nuevoCurso.id) {
-        this.store.dispatch(CursoActions.editarCurso({ curso: this.nuevoCurso }));
+      if (this.isEdit && this.cursoSeleccionado?.id) {
+        // Agregamos manualmente el ID para el PATCH
+        this.store.dispatch(
+          CursoActions.editarCurso({
+            curso: { id: this.cursoSeleccionado.id, ...this.nuevoCurso }
+          })
+        );
       } else {
         this.store.dispatch(CursoActions.agregarCurso({ curso: this.nuevoCurso }));
       }
@@ -44,12 +56,28 @@ export class CursoCreatorComponent implements OnInit {
   }
 
   editarSeleccion(curso: Curso): void {
-    this.nuevoCurso = { ...curso };
+    this.nuevoCurso = { nombre: curso.nombre, descripcion: curso.descripcion };
     this.isEdit = true;
+
+    // Guardamos el ID en una variable separada para la edición
+    this.cursoSeleccionado = curso;
   }
 
   eliminarCurso(id: number): void {
-    this.store.dispatch(CursoActions.eliminarCurso({ id }));
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '300px',
+      data: { mensaje: '¿Estás seguro de que deseas eliminar este curso?' }
+    });
+
+    dialogRef.afterClosed().subscribe((confirmado: boolean) => {
+      if (confirmado) {
+        this.store.dispatch(CursoActions.eliminarCurso({ id }));
+        this.snackBar.open('Curso eliminado correctamente', 'Cerrar', {
+          duration: 3000,
+          panelClass: ['snackbar-success']
+        });
+      }
+    });
   }
 
   resetearFormulario(form: any): void {
